Completion of updateViewCounter function
Realised that the update view count function needs to run when the website is loaded, then the return view count would run
Based on this, consolidated these 2 functions into 1 and deleted returnViewCount function
Function will trigger, update the count or set to 1 if no record exists, then return the new count
