+++ src/functions/updateViewCounter.js
@@ -1,13 +1,65 @@
 const { app } = require('@azure/functions');
+const { TableClient } = require("@azure/data-tables");
+const { json } = require('stream/consumers');
+
+const env = {
+    readOnly: {
+        connectionString: process.env["COSMOSDB_CONNECTION_READONLY"],
+        tableName: "website_visitor_count"
+    },
+    readWrite: {
+        connectionString: process.env["COSMOSDB_CONNECTION_READWRITE"],
+        tableName: "website_visitor_count"
+    }  
+};
+
+const client = {
+    readOnly: TableClient.fromConnectionString(
+        env.readOnly.connectionString,
+        env.readOnly.tableName
+    ),
+    readWrite: TableClient.fromConnectionString(
+        env.readWrite.connectionString,
+        env.readWrite.tableName
+    )
+};
+
+const getCount = () => {
+    return client.readOnly.getEntity("cv-website", "1");
+};
+
+const updateCount = async (count) => {
+    await client.readWrite.upsertEntity(
+        {
+            partitionKey: "cv-website", 
+            rowKey: "1", 
+            count: count
+        }
+    )
+};
 
 app.http('updateViewCounter', {
-    methods: ['GET', 'POST'],
+    methods: ['POST'],
     authLevel: 'anonymous',
     handler: async (request, context) => {
-        context.log(`Http function processed request for url "${request.url}"`);
+        var recordCount = {};
+        var newCount = 0;
 
-        const name = request.query.get('name') || await request.text() || 'world';
+        try {
+            recordCount = await getCount();
+            newCount = recordCount.count + 1;
+            await updateCount(newCount);
+        } catch (error) {
+            if (error.statusCode === 404) {
+                await updateCount(1);
+    
+                recordCount = await getCount();
+            } else {
+                return { body: JSON.stringify(error) };
+            };
+        };        
 
-        return { body: `Hello, ${name}!` };
+        recordCount = await getCount();
+        return { body: recordCount.count };
     }
 });
